Update SDK for iMXRT1170 to Version 2.9.2

+++ middleware/motor_control/pmsm/pmsm_float/mc_identification/pp_measure.c
@@ -28,9 +28,9 @@ void MID_getPp(mid_get_pp_t *sPpMeasFcn)
         sPpMeasFcn->ui16Active                   = TRUE;
         sPpMeasFcn->fltFreqElRamp                = 0.0;
         sPpMeasFcn->ui16PpDetermined             = 0;
-        sPpMeasFcn->sFreqElRampParam.fltRampUp   = sPpMeasFcn->fltFreqElReq / MID_SPEED_RAMP_TIME / 10000;
-        sPpMeasFcn->sFreqElRampParam.fltRampDown = sPpMeasFcn->fltFreqElReq / MID_SPEED_RAMP_TIME / 10000;
-        sPpMeasFcn->sFreqIntegrator.a32Gain      = ACC32(1.0 * sPpMeasFcn->fltFreqMax / 10000.0);
+        sPpMeasFcn->sFreqElRampParam.fltRampUp   = sPpMeasFcn->fltFreqElReq / MID_SPEED_RAMP_TIME / 10000.0F;
+        sPpMeasFcn->sFreqElRampParam.fltRampDown = sPpMeasFcn->fltFreqElReq / MID_SPEED_RAMP_TIME / 10000.0F;
+        sPpMeasFcn->sFreqIntegrator.a32Gain      = ACC32(1.0 * sPpMeasFcn->fltFreqMax / 10000.0F);
         sPpMeasFcn->sFreqIntegrator.f32IAccK_1   = FRAC32(0.0);
         GFLIB_IntegratorInit_F16(0, &sPpMeasFcn->sFreqIntegrator);
         GFLIB_RampInit_FLT(0, &sPpMeasFcn->sFreqElRampParam);
@@ -40,7 +40,7 @@ void MID_getPp(mid_get_pp_t *sPpMeasFcn)
     *(sPpMeasFcn->pfltIdReq) = sPpMeasFcn->fltIdReqOpenLoop;
 
     /* Else start incrementing position */
-    if (sPpMeasFcn->ui16WaitingSteady == 0)
+    if (sPpMeasFcn->ui16WaitingSteady == 0U)
     {
         /* Ramp electrical speed */
         sPpMeasFcn->fltFreqElRamp = GFLIB_Ramp_FLT(sPpMeasFcn->fltFreqElReq, &sPpMeasFcn->sFreqElRampParam);
@@ -51,7 +51,7 @@ void MID_getPp(mid_get_pp_t *sPpMeasFcn)
 
     /* If position overflows, wait 2400ms in zero position */
     if (((*sPpMeasFcn->pf16PosEl < FRAC16(0.0)) && (sPpMeasFcn->f16PosElLast > FRAC16(0.0))) ||
-        (sPpMeasFcn->ui16WaitingSteady == 1))
+        (sPpMeasFcn->ui16WaitingSteady == 1U))
     {
         *sPpMeasFcn->pf16PosEl = FRAC16(-1.0);
 
@@ -65,7 +65,7 @@ void MID_getPp(mid_get_pp_t *sPpMeasFcn)
         sPpMeasFcn->ui16LoopCounter++;
 
         /* Escape waiting in steady position after 2400 ms */
-        if (sPpMeasFcn->ui16LoopCounter > MID_TIME_2400MS)
+        if (sPpMeasFcn->ui16LoopCounter > (uint16_t)MID_TIME_2400MS)
         {
             *sPpMeasFcn->pf16PosEl        = FRAC16(0.0);
             sPpMeasFcn->f16PosElLast      = FRAC16(0.0);
@@ -77,7 +77,7 @@ void MID_getPp(mid_get_pp_t *sPpMeasFcn)
     sPpMeasFcn->f16PosElLast    = sPpMeasFcn->f16PosElCurrent;
     sPpMeasFcn->f16PosElCurrent = *sPpMeasFcn->pf16PosEl;
 
-    if (sPpMeasFcn->ui16PpDetermined > 0)
+    if (sPpMeasFcn->ui16PpDetermined > 0U)
     {
         /* When finished exit the function */
         sPpMeasFcn->ui16Active = FALSE;
